requirements.txt
asgiref==3.3.4; python_version >= "3.6"
certifi==2020.12.5; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.5"
chardet==4.0.0; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.5"
configparser==5.0.2; python_version >= "3.6"
dj-database-url==0.5.0
django-appconf==1.0.4
django-cors-headers==3.7.0; python_version >= "3.6"
django-enumfield==2.0.2
django-filter==2.4.0; python_version >= "3.5"
django-haystack==3.0
django-heroku==0.3.1
django-measurement==3.2.3
django==3.2.3; python_version >= "3.6"
djangorestframework-simplejwt==4.7.1; python_version >= "3.7"
djangorestframework==3.12.4; python_version >= "3.5"
drf-haystack==1.8.10
gunicorn==20.1.0; python_version >= "3.5"
herepy==3.5.2; python_version >= "3.5"
idna==2.10; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.5"
markdown==3.3.4; python_version >= "3.6"
measurement==3.2.0
mpmath==1.2.1; python_version >= "3.6"
psycopg2-binary==2.9.1; python_version >= "3.6"
psycopg2==2.9.1; python_version >= "3.6"
pydantic==1.8.2; python_full_version >= "3.6.1"
pyjavaproperties==0.7
pyjwt==2.1.0; python_version >= "3.7"
pymemcache==3.4.4
python-dateutil==2.8.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
python-memcached==1.59
pytz==2021.1; python_version >= "3.6"
requests==2.25.1; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.5"
serializer==0.2.1
simplejson==3.17.2; python_version >= "2.5" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
six==1.16.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
sqlparse==0.4.1; python_version >= "3.6"
sympy==1.8; python_version >= "3.6"
typing-extensions==3.10.0.0; python_full_version >= "3.6.1"
urllib3==1.26.5; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.5"
whitenoise==5.2.0; python_version >= "3.5" and python_version < "4"
